+++ examples/chinook/sqlite/load.sql
@@ -1,0 +1,13 @@
+.mode csv
+.header on
+.import ../csv/albums.csv albums
+.import ../csv/artists.csv artists
+.import ../csv/customers.csv customers
+.import ../csv/employees.csv employees
+.import ../csv/genres.csv genres
+.import ../csv/invoice_items.csv invoice_items
+.import ../csv/invoices.csv invoices
+.import ../csv/media_types.csv media_types
+.import ../csv/playlists.csv playlists
+.import ../csv/playlist_track.csv playlist_track
+.import ../csv/tracks.csv tracks